chore: fix build errors
