--- bciers/libs/shared/types/src/index.ts
@@ -1,7 +1,0 @@
-export type ContentItem = {
-  icon: string;
-  href: string;
-  title: string;
-  content: string;
-  links?: { title?: string; href: string; notification?: string }[];
-};
