+++ bciers/libs/shared/types/src/session.ts
@@ -1,0 +1,8 @@
+export type Session = {
+  user?: {
+    full_name?: string;
+    given_name?: string;
+    family_name?: string;
+    name?: string; // Added this for the fallback case
+  };
+};
