+++ bciers/libs/shared/types/src/tiles.ts
@@ -1,0 +1,7 @@
+export type ContentItem = {
+  icon: string;
+  href: string;
+  title: string;
+  content: string;
+  links?: { title?: string; href: string; notification?: string }[];
+};
