+++ bciers/tsconfig.base.json
@@ -29,7 +29,7 @@
       "@bciers/styles": ["libs/shared/styles/src/index.ts"],
       "@bciers/styles/*": ["libs/shared/styles/src/lib/*"],
       "@bciers/testConfig/*": ["libs/shared/testConfig/src/*"],
-      "@bciers/types": ["libs/shared/types/src/index.ts"],
+      "@bciers/types/*": ["libs/shared/types/src/*"],
       "@bciers/utils/*": ["libs/shared/utils/*"],
       "@bciers/utils/server": ["libs/shared/utils/server.ts"]
     }
